requirements.txt
fastapi[standard]
uvicorn>=0.24.0
sqlalchemy>=2.0.23
pydantic>=2.4.2
python-dotenv>=1.0.0
sqlmodel>=0.0.14
groq
aiohttp
feedparser